--- src/core/main/index.ts
@@ -1,1 +1,0 @@
-export type { ICorePlaybackObservation } from "./worker/worker_playback_observer";
