+++ src/core/stream/period/utils/get_adaptation_switch_strategy.ts
@@ -15,8 +15,8 @@
  */
 
 import config from "../../../../config";
-import type { IReadOnlyPlaybackObserver } from "../../../../main_thread/types";
 import type { IAdaptation, IPeriod } from "../../../../manifest";
+import type { IReadOnlyPlaybackObserver } from "../../../../playback_observer";
 import areCodecsCompatible from "../../../../utils/are_codecs_compatible";
 import type { IRange } from "../../../../utils/ranges";
 import {
