+++ src/core/stream/representation/utils/get_buffer_status.ts
@@ -15,13 +15,13 @@
  */
 
 import config from "../../../../config";
-import type { IReadOnlyPlaybackObserver } from "../../../../main_thread/types";
 import type {
   IManifest,
   IAdaptation,
   IPeriod,
   IRepresentation,
 } from "../../../../manifest";
+import type { IReadOnlyPlaybackObserver } from "../../../../playback_observer";
 import isNullOrUndefined from "../../../../utils/is_null_or_undefined";
 import type {
   IBufferedChunk,
