+++ src/core/stream/representation/utils/push_media_segment.ts
@@ -15,7 +15,6 @@
  */
 
 import config from "../../../../config";
-import type { IReadOnlyPlaybackObserver } from "../../../../main_thread/types";
 import type {
   IManifest,
   IAdaptation,
@@ -23,6 +22,7 @@ import type {
   IPeriod,
   IRepresentation,
 } from "../../../../manifest";
+import type { IReadOnlyPlaybackObserver } from "../../../../playback_observer";
 import type { ISegmentParserParsedMediaChunk } from "../../../../transports";
 import objectAssign from "../../../../utils/object_assign";
 import type { IReadOnlySharedReference } from "../../../../utils/reference";
