+++ src/main_thread/api/index.ts
@@ -14,15 +14,5 @@
  * limitations under the License.
  */
 
-import PlaybackObserver from "./playback_observer";
 import Player from "./public_api";
-export { PlaybackObserver };
-export {
-  IObservationPosition,
-  IPlaybackObservation,
-  IPlaybackObserverEventType,
-  IReadOnlyPlaybackObserver,
-  IFreezingStatus,
-  IRebufferingStatus,
-} from "./playback_observer";
 export default Player;
