+++ src/main_thread/types.ts
@@ -1,11 +1,5 @@
 import type RxPlayer from "./api";
 import type {
-  IFreezingStatus,
-  IRebufferingStatus,
-  IObservationPosition,
-  IReadOnlyPlaybackObserver,
-} from "./api";
-import type {
   IContentProtection,
   IProcessedProtectionData,
 } from "./decrypt";
@@ -17,12 +11,6 @@ import type {
 export type IRxPlayer = RxPlayer;
 
 export type {
-  // Playback Observation Metadata
-  IObservationPosition,
-  IReadOnlyPlaybackObserver,
-  IFreezingStatus,
-  IRebufferingStatus,
-
   // Decrypt Metadata
   IContentProtection,
   IProcessedProtectionData,
