+++ src/playback_observer/media_element_playback_observer.ts
@@ -14,21 +14,28 @@
  * limitations under the License.
  */
 
-import isSeekingApproximate from "../../compat/is_seeking_approximate";
-import config from "../../config";
-import log from "../../log";
-import getMonotonicTimeStamp from "../../utils/monotonic_timestamp";
-import noop from "../../utils/noop";
-import objectAssign from "../../utils/object_assign";
-import { getBufferedTimeRange } from "../../utils/ranges";
+import isSeekingApproximate from "../compat/is_seeking_approximate";
+import config from "../config";
+import log from "../log";
+import getMonotonicTimeStamp from "../utils/monotonic_timestamp";
+import noop from "../utils/noop";
+import objectAssign from "../utils/object_assign";
+import { getBufferedTimeRange } from "../utils/ranges";
+import type { IReadOnlySharedReference } from "../utils/reference";
+import SharedReference from "../utils/reference";
+import type { CancellationSignal } from "../utils/task_canceller";
+import TaskCanceller from "../utils/task_canceller";
 import type {
-  IReadOnlySharedReference,
-} from "../../utils/reference";
-import SharedReference from "../../utils/reference";
-import type {
-  CancellationSignal,
-} from "../../utils/task_canceller";
-import TaskCanceller from "../../utils/task_canceller";
+  IMediaInfos,
+  IPlaybackObservation,
+  IPlaybackObserverEventType,
+  IReadOnlyPlaybackObserver,
+  IRebufferingStatus,
+  IFreezingStatus,
+} from "./types";
+import { SeekingState } from "./types";
+import generateReadOnlyObserver from "./utils/generate_read_only_observer";
+import ObservationPosition from "./utils/observation_position";
 
 /**
  * HTMLMediaElement Events for which playback observations are calculated and
@@ -503,352 +510,6 @@ export default class PlaybackObserver {
   }
 }
 
-/** "Event" that triggered the playback observation. */
-export type IPlaybackObserverEventType =
-  /** First playback observation automatically emitted. */
-  "init" |
-  /** Observation manually forced by the PlaybackObserver. */
-  "manual" |
-  /** Regularly emitted playback observation when no event happened in a long time. */
-  "timeupdate" |
-  /** On the HTML5 event with the same name */
-  "canplay" |
-  /** On the HTML5 event with the same name */
-  "ended" |
-  /** On the HTML5 event with the same name */
-  "canplaythrough" | // HTML5 Event
-  /** On the HTML5 event with the same name */
-  "play" |
-  /** On the HTML5 event with the same name */
-  "pause" |
-  /** On the HTML5 event with the same name */
-  "seeking" |
-  /** On the HTML5 event with the same name */
-  "seeked" |
-  /** On the HTML5 event with the same name */
-  "stalled" |
-  /** On the HTML5 event with the same name */
-  "loadedmetadata" |
-  /** On the HTML5 event with the same name */
-  "ratechange" |
-  /** An internal seek happens */
-  "internal-seeking";
-
-/** Information recuperated on the media element on each playback observation. */
-interface IMediaInfos {
-  /** Value of `buffered` (buffered ranges) for the media element. */
-  buffered : TimeRanges;
-  /**
-   * `currentTime` (position) set on the media element at the time of the
-   * PlaybackObserver's measure.
-   */
-  position : number;
-  /** Current `duration` set on the media element. */
-  duration : number;
-  /** Current `ended` set on the media element. */
-  ended: boolean;
-  /** Current `paused` set on the media element. */
-  paused : boolean;
-  /** Current `playbackRate` set on the media element. */
-  playbackRate : number;
-  /** Current `readyState` value on the media element. */
-  readyState : number;
-  /** Current `seeking` value on the mediaElement. */
-  seeking : boolean;
-}
-
-/** Categorize a pending seek operation. */
-export const enum SeekingState {
-  /** We're not currently seeking. */
-  None,
-  /**
-   * We're currently seeking due to an internal logic of the RxPlayer (e.g.
-   * discontinuity skipping).
-   */
-  Internal,
-  /** We're currently seeking due to a regular seek wanted by the application. */
-  External,
-}
-
-/**
- * Describes when the player is "rebuffering" and what event started that
- * status.
- * "Rebuffering" is a status where the player has not enough buffer ahead to
- * play reliably.
- * The RxPlayer should pause playback when a playback observation indicates the
- * rebuffering status.
- */
-export interface IRebufferingStatus {
-  /** What started the player to rebuffer. */
-  reason : "seeking" | // Building buffer after seeking
-           "not-ready" | // Building buffer after low readyState
-           "buffering"; // Other cases
-  /**
-   * Monotonically-raising timestamp at the time the rebuffering happened on the
-   * main thread.
-   */
-  timestamp : number;
-  /**
-   * Position, in seconds, at which data is awaited.
-   * If `null` the player is rebuffering but not because it is awaiting future data.
-   * If `undefined`, that position is unknown.
-   */
-  position : number | null | undefined;
-}
-
-/**
- * Describes when the player is "frozen".
- * This status is reserved for when the player is stuck at the same position for
- * an unknown reason.
- */
-export interface IFreezingStatus {
-  /**
-   * Monotonically-raising timestamp at the time the freezing started to be
-   * detected.
-   */
-  timestamp : number;
-}
-
-/** Information emitted on each playback observation. */
-export interface IPlaybackObservation extends Omit<IMediaInfos, "position" | "seeking"> {
-   /** Event that triggered this playback observation. */
-  event : IPlaybackObserverEventType;
-  /** Current seeking state. */
-  seeking : SeekingState;
-  /**
-   * Information on the current position being played, the position for which
-   * media is wanted etc.
-   */
-  position : IObservationPosition;
-  /**
-   * Set if the player is short on audio and/or video media data and is a such,
-   * rebuffering.
-   * `null` if not.
-   */
-  rebuffering : IRebufferingStatus | null;
-  /**
-   * Set if the player is frozen, that is, stuck in place for unknown reason.
-   * Note that this reason can be a valid one, such as a necessary license not
-   * being obtained yet.
-   *
-   * `null` if the player is not frozen.
-   */
-  freezing : IFreezingStatus | null;
-  /**
-   * Gap between `currentTime` and the next position with un-buffered data.
-   * `Infinity` if we don't have buffered data right now.
-   * `undefined` if we cannot determine the buffer gap.
-   */
-  bufferGap : number | undefined;
-  /**
-   * The buffered range we are currently playing.
-   * `null` if no range is currently available.
-   * `undefined` if we cannot tell which range is currently available.
-   */
-  currentRange : { start : number;
-                   end : number; } |
-                 null |
-                 undefined;
-}
-
-export type IObservationPosition = ObservationPosition;
-
-export class ObservationPosition {
-  /**
-   * Known position at the time the Observation was emitted, in seconds.
-   *
-   * Note that it might have changed since. If you want truly precize
-   * information, you should recuperate it from the HTMLMediaElement directly
-   * through another mean.
-   */
-  private _last: number;
-  /**
-   * Actually wanted position in seconds that is not yet reached.
-   *
-   * This might for example be set to the initial position when the content is
-   * loading (and thus potentially at a `0` position) but which will be seeked
-   * to a given position once possible. It may also be the position of a seek
-   * that has not been properly accounted for by the current device.
-   */
-  private _wanted: number | null;
-  constructor(last: number, wanted: number | null) {
-    this._last = last;
-    this._wanted = wanted;
-  }
-
-  /**
-   * Obtain arguments allowing to instanciate the same ObservationPosition.
-   *
-   * This can be used to create a new `ObservationPosition` across JS realms,
-   * generally to communicate its data between the main thread and a WebWorker.
-   * @returns {Array.<number>}
-   */
-  public serialize(): [number, number | null] {
-    return [this._last, this._wanted];
-  }
-
-  /**
-   * Returns the playback position actually observed on the media element at
-   * the time the playback observation was made.
-   *
-   * Note that it may be different than the position for which media data is
-   * wanted in rare scenarios where the goal position is not yet set on the
-   * media element.
-   *
-   * You should use this value when you want to obtain the actual position set
-   * on the media element for browser compatibility purposes. Note that this
-   * position was calculated at observation time, it might thus not be
-   * up-to-date if what you want is milliseconds-accuracy.
-   *
-   * If what you want is the actual position which the player is intended to
-   * play, you should rely on `getWanted` instead`.
-   * @returns {number}
-   */
-  public getPolled(): number {
-    return this._last;
-  }
-
-  /**
-   * Returns the position which the player should consider to load media data
-   * at the time the observation was made.
-   *
-   * It can be different than the value returned by `getPolled` in rare
-   * scenarios:
-   *
-   *   - When the initial position has not been set yet.
-   *
-   *   - When the current device do not let the RxPlayer peform precize seeks,
-   *     usually for perfomance reasons by seeking to a previous IDR frame
-   *     instead (for now only Tizen may be like this), in which case we
-   *     prefer to generally rely on the position wanted by the player (this
-   *     e.g. prevents issues where the RxPlayer logic and the device are
-   *     seeking back and forth in a loop).
-   *
-   *   - When a wanted position has been "forced" (@see forceWantedPosition).
-   * @returns {number}
-   */
-  public getWanted(): number {
-    return this._wanted ?? this._last;
-  }
-
-  /**
-   * Method to call if you want to overwrite the currently wanted position.
-   * @param {number} pos
-   */
-  public forceWantedPosition(pos: number): void {
-    this._wanted = pos;
-  }
-
-  /**
-   * Returns `true` when the position wanted returned by `getWanted` and the
-   * actual position returned by `getPolled` may be different, meaning that
-   * we're currently not at the position we want to reach.
-   *
-   * This is a relatively rare situation which only happens when either the
-   * initial seek has not yet been performed. on specific targets where the
-   * seeking behavior is a little broken (@see getWanted) or when the wanted
-   * position has been forced (@see forceWantedPosition).
-   *
-   * In those situations, you might temporarily refrain from acting upon the
-   * actual current media position, as it may change soon.
-   *
-   * @returns {boolean}
-   */
-  public isAwaitingFuturePosition(): boolean {
-    return this._wanted !== null;
-  }
-}
-
-/**
- * Interface providing a generic and read-only version of a `PlaybackObserver`.
- *
- * This interface allows to provide regular and specific playback information
- * without allowing any effect on playback like seeking.
- *
- * This can be very useful to give specific playback information to modules you
- * don't want to be able to update playback.
- *
- * Note that a `PlaybackObserver` is compatible and can thus be upcasted to a
- * `IReadOnlyPlaybackObserver` to "remove" its right to update playback.
- */
-export interface IReadOnlyPlaybackObserver<TObservationType> {
-  /**
-   * Get the current playing position, in seconds.
-   * Returns `undefined` when this cannot be known, such as when the playback
-   * observer is running in a WebWorker.
-   * @returns {number|undefined}
-   */
-  getCurrentTime() : number | undefined;
-  /**
-   * Returns the current playback rate advertised by the `HTMLMediaElement`.
-   * Returns `undefined` when this cannot be known, such as when the playback
-   * observer is running in a WebWorker.
-   * @returns {number|undefined}
-   */
-  getPlaybackRate() : number | undefined;
-  /**
-   * Get the HTMLMediaElement's current `readyState`.
-   * Returns `undefined` when this cannot be known, such as when the playback
-   * observer is running in a WebWorker.
-   * @returns {number|undefined}
-   */
-  getReadyState() : number | undefined;
-  /**
-   * Returns the current `paused` status advertised by the `HTMLMediaElement`.
-   *
-   * Use this instead of the same status emitted on an observation when you want
-   * to be sure you're using the current value.
-   *
-   * Returns `undefined` when this cannot be known, such as when the playback
-   * observer is running in a WebWorker.
-   * @returns {boolean|undefined}
-   */
-  getIsPaused() : boolean | undefined;
-  /**
-   * Returns an `IReadOnlySharedReference` storing the last playback observation
-   * produced by the `IReadOnlyPlaybackObserver` and updated each time a new one
-   * is produced.
-   *
-   * This value can then be for example listened to to be notified of future
-   * playback observations.
-   *
-   * @returns {Object}
-   */
-  getReference() : IReadOnlySharedReference<TObservationType>;
-  /**
-   * Register a callback so it regularly receives playback observations.
-   * @param {Function} cb
-   * @param {Object} options - Configuration options:
-   *   - `includeLastObservation`: If set to `true` the last observation will
-   *     be first emitted synchronously.
-   *   - `clearSignal`: If set, the callback will be unregistered when this
-   *     CancellationSignal emits.
-   * @returns {Function} - Allows to easily unregister the callback
-   */
-  listen(
-    cb : (
-      observation : TObservationType,
-      stopListening : () => void
-    ) => void,
-    options? : { includeLastObservation? : boolean | undefined;
-                 clearSignal? : CancellationSignal | undefined; }
-  ) : void;
-  /**
-   * Generate a new `IReadOnlyPlaybackObserver` from this one.
-   *
-   * As argument, this method takes a function which will allow to produce
-   * the new set of properties to be present on each observation.
-   * @param {Function} transform
-   * @returns {Object}
-   */
-  deriveReadOnlyObserver<TDest>(
-    transform : (
-      observationRef : IReadOnlySharedReference<TObservationType>,
-      cancellationSignal : CancellationSignal
-    ) => IReadOnlySharedReference<TDest>
-  ) : IReadOnlyPlaybackObserver<TDest>;
-}
-
 /**
  * Returns the amount of time in seconds the buffer should have ahead of the
  * current position before resuming playback. Based on the infos of the
@@ -1217,67 +878,6 @@ function prettyPrintBuffered(
     currentTimeStr = " ".repeat(str.length) + `^${currentTime}`;
   }
   return str + "\n" + currentTimeStr;
-}
-
-/**
- * Create `IReadOnlyPlaybackObserver` from a source `IReadOnlyPlaybackObserver`
- * and a mapping function.
- * @param {Object} src
- * @param {Function} transform
- * @returns {Object}
- */
-export function generateReadOnlyObserver<TSource, TDest>(
-  src : IReadOnlyPlaybackObserver<TSource>,
-  transform : (
-    observationRef : IReadOnlySharedReference<TSource>,
-    cancellationSignal : CancellationSignal
-  ) => IReadOnlySharedReference<TDest>,
-  cancellationSignal : CancellationSignal
-) : IReadOnlyPlaybackObserver<TDest> {
-  const mappedRef = transform(src.getReference(), cancellationSignal);
-  return {
-    getCurrentTime() {
-      return src.getCurrentTime();
-    },
-    getReadyState() {
-      return src.getReadyState();
-    },
-    getPlaybackRate() {
-      return src.getPlaybackRate();
-    },
-    getIsPaused() {
-      return src.getIsPaused();
-    },
-    getReference() : IReadOnlySharedReference<TDest> {
-      return mappedRef;
-    },
-    listen(
-      cb : (
-        observation : TDest,
-        stopListening : () => void
-      ) => void,
-      options? : { includeLastObservation? : boolean | undefined;
-                   clearSignal? : CancellationSignal | undefined; }
-    ) : void {
-      if (cancellationSignal.isCancelled() ||
-          options?.clearSignal?.isCancelled() === true)
-      {
-        return ;
-      }
-      mappedRef.onUpdate(cb, {
-        clearSignal: options?.clearSignal,
-        emitCurrentValue: options?.includeLastObservation,
-      });
-    },
-    deriveReadOnlyObserver<TNext>(
-      newTransformFn : (
-        observationRef : IReadOnlySharedReference<TDest>,
-        signal : CancellationSignal
-      ) => IReadOnlySharedReference<TNext>
-    ) : IReadOnlyPlaybackObserver<TNext> {
-      return generateReadOnlyObserver(this, newTransformFn, cancellationSignal);
-    },
-  };
 }
 
 /**
